rename: useCalculateInnerDisplayStyle -> useCalculateInnerPcStyle #2
