+++ src/hooks/useCalculateFirstViewAnimatedSize.ts
@@ -2,7 +2,7 @@ import { useRef, useMemo, useCallback, useEffect } from 'react'
 import gsap from 'gsap'
 import ScrollTrigger from 'gsap/ScrollTrigger'
 import { useWatchInnerAspect } from 'hooks/useWatchInnerAspect'
-import { useCalculateInnerPcStyle } from 'hooks/useCalculateInnerDisplayStyle'
+import { useCalculateInnerPcStyle } from 'hooks/useCalculateInnerPcStyle'
 
 /**
  * ファーストビューでスクロールした時に画面内のPCがピッタリ実際の画面に収まるような拡大率・位置を計算し、アニメーションを付与する
